[DEV-2206] replace migration
